feat(dashboard/sidebar): move code to components
