+++ packages/components/_components-themes.scss
@@ -4,12 +4,12 @@
 @import 'placeholder/block-placeholder/xm-block-placeholder-theme';
 @import 'text/xm-text-theme';
 @import 'code/xm-code-theme';
-@import 'packages/dashboard/menu/menu-theme';
-@import 'packages/dashboard/sidebar-user/user-theme';
+@import 'menu/menu-theme';
+@import 'sidebar-user/user-theme';
 @import 'ribbon/xm-ribbon-theme';
 @import 'multilanguage/xm-angular-editor-theme';
 @import '../entity/search/xm-navbar-search-widget-theme';
-@import 'packages/dashboard/sidebar/sidebar-theme';
+@import 'sidebar/sidebar-theme';
 @import 'table/table-widget/xm-table-widget-theme';
 
 @mixin xm-components-theme($xm-theme) {
