+++ packages/components/menu/ng-package.json
