+++ packages/components/navbar-user-widget/navbar-user-widget.component.spec.ts
