+++ packages/components/navbar-user-widget/ng-package.json
