+++ packages/components/sidebar-user/_user-theme.scss
