+++ packages/components/sidebar-user/ng-package.json
