+++ packages/components/sidebar-user/sidebar-user-subtitle.ts
@@ -3,7 +3,7 @@ import { XmUser } from '@xm-ngx/core/user';
 import { Translate, XmTranslationModule } from '@xm-ngx/translation';
 import { get } from 'lodash';
 import { CommonModule } from '@angular/common';
-import { MenuComponent } from '@xm-ngx/dashboard/menu';
+import { MenuComponent } from '@xm-ngx/components/menu';
 import { XmPermissionModule } from '@xm-ngx/core/permission';
 import { XmDynamicModule } from '@xm-ngx/dynamic';
