+++ packages/components/sidebar-user/sidebar-user.component.html
