+++ packages/components/sidebar-user/sidebar-user.component.scss
