+++ packages/components/sidebar-user/sidebar-user.component.spec.ts
