+++ packages/components/sidebar-user/sidebar-user.component.ts
@@ -1,19 +1,19 @@
 import { Component } from '@angular/core';
 import { matExpansionAnimations } from '@angular/material/expansion';
 import * as _ from 'lodash';
-import { MenuItem } from '@xm-ngx/dashboard/menu';
+import { MenuItem } from '@xm-ngx/components/menu';
 import { UserWidgetBase } from './user-widget.base';
 import { DashboardStore } from '@xm-ngx/core/dashboard';
 import { XmUserService } from '@xm-ngx/core/user';
 import { ContextService } from '@xm-ngx/core/context';
 import { Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { MenuComponent } from '@xm-ngx/dashboard/menu';
+import { MenuComponent } from '@xm-ngx/components/menu';
 import { XmPermissionModule } from '@xm-ngx/core/permission';
 import { XmTranslationModule } from '@xm-ngx/translation';
 import { XmDynamicModule } from '@xm-ngx/dynamic';
 import { SidebarUserSubtitle } from './sidebar-user-subtitle';
-import { MenuLinkComponent } from '@xm-ngx/dashboard/menu';
+import { MenuLinkComponent } from '@xm-ngx/components/menu';
 
 
 @Component({
