+++ packages/components/sidebar/_sidebar-theme.scss
