+++ packages/components/sidebar/configs/xm-sidebar.config.ts
