+++ packages/components/sidebar/ng-package.json
