+++ packages/components/sidebar/stores/xm-sidebar-store.service.ts
