+++ packages/components/sidebar/xm-sidebar-state.component.scss
