+++ packages/dashboard/module/xm-ngx-dashboard.module.ts
@@ -5,16 +5,8 @@ import { XmDynamicModule } from '@xm-ngx/dynamic';
     imports: [
         XmDynamicModule.forChild([
             {
-                selector: '@xm-ngx/dashboard/default-dashboard',
+                selector: 'default-dashboard',
                 loadChildren: () => import('@xm-ngx/dashboard').then(m => m.XmDashboardModule),
-            },
-            {
-                selector: '@xm-ngx/components/sidebar-menu',
-                loadChildren: () => import('@xm-ngx/dashboard/menu').then(m => m.MenuComponent),
-            },
-            {
-                selector: '@xm-ngx/components/navbar-user-widget',
-                loadChildren: () => import('@xm-ngx/dashboard/navbar-user-widget').then(m => m.NavbarUserWidgetComponent),
             },
         ]),
     ],
