+++ packages/operators/index.ts
@@ -19,3 +19,4 @@ export { uuid } from './src/uuid';
 export { Interpolate, interpolate } from './src/interpolate';
 export { randomInt } from './src/random-int';
 export * from './src/request-util';
+export * from './src/transpilingForIE';
